+++ AddOns/!KalielsTracker/Modules/Addon_PetTracker.lua
@@ -318,7 +318,7 @@ function M:OnInitialize()
 	_DBG("|cffffff00Init|r - "..self:GetName(), true)
 	db = KT.db.profile
 	dbChar = KT.db.char
-    self.isAvailable = (KT:CheckAddOn("PetTracker", "11.2.3") and db.addonPetTracker)
+    self.isAvailable = (KT:CheckAddOn("PetTracker", "11.2.7") and db.addonPetTracker)
 
 	if self.isAvailable then
 		KT:Alert_IncompatibleAddon("PetTracker", "11.1.10")